Create Logon Script
I wrote this as a logon script that would map users to their network shared drives as well as networked printers. In full use it is slightly more detailed and has more customization for the individuals that have unique needs but this is the universal source code that I am allowed to share.